# Optimise `ImageMode.getmode` using `functools.lru_cache`